+++ src/anyio/__init__.py
@@ -2,17 +2,16 @@ __all__ = ('run', 'sleep', 'current_time', 'get_all_backends', 'get_cancelled_ex
            'BrokenResourceError', 'BusyResourceError', 'ClosedResourceError', 'DelimiterNotFound',
            'EndOfStream', 'ExceptionGroup', 'IncompleteRead', 'WouldBlock', 'AsyncFile',
            'open_file', 'aclose_forcefully', 'open_signal_receiver', 'connect_tcp',
-           'connect_tcp_with_tls', 'connect_unix', 'create_tcp_listener', 'create_unix_listener',
-           'create_udp_socket', 'create_connected_udp_socket', 'getaddrinfo', 'getnameinfo',
-           'wait_socket_readable', 'wait_socket_writable', 'create_memory_object_stream',
-           'run_process', 'open_process', 'create_lock', 'create_condition', 'create_event',
-           'create_semaphore', 'create_capacity_limiter', 'open_cancel_scope', 'fail_after',
-           'move_on_after', 'current_effective_deadline', 'create_task_group', 'TaskInfo',
-           'get_current_task', 'get_running_tasks', 'wait_all_tasks_blocked',
-           'run_sync_in_worker_thread', 'run_async_from_thread',
-           'current_default_worker_thread_limiter', 'create_blocking_portal',
-           'start_blocking_portal', 'typed_attribute', 'TypedAttributeSet',
-           'TypedAttributeProvider')
+           'connect_unix', 'create_tcp_listener', 'create_unix_listener', 'create_udp_socket',
+           'create_connected_udp_socket', 'getaddrinfo', 'getnameinfo', 'wait_socket_readable',
+           'wait_socket_writable', 'create_memory_object_stream', 'run_process', 'open_process',
+           'create_lock', 'create_condition', 'create_event', 'create_semaphore',
+           'create_capacity_limiter', 'open_cancel_scope', 'fail_after', 'move_on_after',
+           'current_effective_deadline', 'create_task_group', 'TaskInfo', 'get_current_task',
+           'get_running_tasks', 'wait_all_tasks_blocked', 'run_sync_in_worker_thread',
+           'run_async_from_thread', 'current_default_worker_thread_limiter',
+           'create_blocking_portal', 'start_blocking_portal', 'typed_attribute',
+           'TypedAttributeSet', 'TypedAttributeProvider')
 
 from ._core._eventloop import current_time, get_all_backends, get_cancelled_exc_class, run, sleep
 from ._core._exceptions import (
@@ -22,9 +21,8 @@ from ._core._fileio import AsyncFile, open_file
 from ._core._resources import aclose_forcefully
 from ._core._signals import open_signal_receiver
 from ._core._sockets import (
-    connect_tcp, connect_tcp_with_tls, connect_unix, create_connected_udp_socket,
-    create_tcp_listener, create_udp_socket, create_unix_listener, getaddrinfo, getnameinfo,
-    wait_socket_readable, wait_socket_writable)
+    connect_tcp, connect_unix, create_connected_udp_socket, create_tcp_listener, create_udp_socket,
+    create_unix_listener, getaddrinfo, getnameinfo, wait_socket_readable, wait_socket_writable)
 from ._core._streams import create_memory_object_stream
 from ._core._subprocesses import open_process, run_process
 from ._core._synchronization import (