+++ docs/versionhistory.rst
@@ -8,8 +8,9 @@ This library adheres to `Semantic Versioning 2.0 <http://semver.org/>`_.
 - **BACKWARDS INCOMPATIBLE** Replaced individual stream/listener attributes with a system of
   typed extra attributes which traverses the entire chain of wrapped streams/listeners to find the
   attribute being looked for
-- Added the ``tls_standard_compatible`` argument to ``connect_tcp_with_tls()`` (this was mistakenly
-  omitted when ``connect_tcp()`` was split in two)
+- **BACKWARDS INCOMPATIBLE** Merged ``connect_tcp_with_tls()`` back into ``connect_tcp()``
+- Added the ``tls_standard_compatible`` argument to ``connect_tcp()`` (this was mistakenly omitted
+  when ``connect_tcp()`` was split in two)
 - Fixed Hypothesis support in the pytest plugin (it was not actually running the Hypothesis tests
   at all)